[WFLY-11406] Properly fix DefaultClassFactory so it can be used on JDK12 and above.
Some background about this change:

After bisecting openjdk code base we identified this commit
https://github.com/openjdk/jdk/commit/3f0daceac06ec44186130874401a99378d076768
which introduced new JDK behaviour WRT reflective access to some fields and methods.
Since this commit it is impossible to:

A) Reflectively access all fields and methods of:

    Reflection.class
    AccessibleObject.class
    ClassLoader.class
    Constructor.class
    Field.class
    Method.class

B) Reflectively access fields:

    Class.class -> "classLoader"
    System.class -> "security"

The proper fix is to use MethodHandles to workaround this reflection limitations.